backend/requirements.txt
# ADK and AI dependencies
google-generativeai
google-cloud-aiplatform
# adk-python # Uncomment when available or use specific source
python-dotenv
fastapi
uvicorn
pydantic-ai
pydantic-graph

